cleverbots/requirements.txt
certifi==2018.8.13
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
defusedxml==0.5.0
Django==2.2
django-bootstrap4==0.0.8
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
gevent==1.4.0
grab==0.6.41
greenlet==0.4.15
gunicorn==19.9.0
httpie==1.0.2
idna==2.7
oauthlib==3.0.1
Pillow==5.3.0
pip-tools==3.4.0
psycopg2==2.7.7
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
python3-openid==3.1.0
pytils==0.3
pytz==2018.7
requests==2.19.1
requests-oauthlib==1.2.0
selection==0.0.14
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sql==0.4.0
sqlparse==0.3.0
urllib3==1.23
user-agent==0.1.9
weblib==0.1.30
